Fix `guarantee_final_pipeline` feature
Background:
There are cases when a project needs to have a final run of the CI
pipeline before the merge.
For example, if a team practices fixup-driven code review process in a
project, the fixup commits must be auto-squashed prior to a merge. This
can be done in CI check by having a job that performs the check only
when Marge-bot is assigned. Gitlab CI is not triggered automatically on
actions like assigning, so that Marge-bot `--guarantee-final-pipeline`
option is what is exactly needed in this case: it is supposed to trigger
a pipeline right before the merge.

Problem and cause:
`--guarantee-final-pipeline` does not trigger Gitlab CI at all. Only a
comment `jenkins retry` appears in the review. Looking to the code, it
is visible that no more than additional 30 seconds of waiting is
implemented.

Measure:
- When `--guarantee-final-pipeline` is specified, change the Marge-bot
  behavior to triggering a merge request pipeline (falling back to a
  regular branch pipeline if [MR workflow][0] is not set up) and waiting
  for it to pass.
- Check the solution in Marge-bot tests.

[0]: https://docs.gitlab.com/ee/ci/pipelines/merge_request_pipelines.html#prerequisites