Add pipeline reference to Marge's messages
Make it easier to identify the pipeline which caused Marge to give a
certain reply by adding the pipeline on the message.

This also facilitates large projects to run scripts to raise statistics
on the CI, identify the most common reasons why it fails and quick
solutions.

See https://gitlab.com/marge-org/marge-bot/-/issues/380

Signed-off-by: Helen Koike <helen.koike@collabora.com>